% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABDS.R
\name{ABDS}
\alias{ABDS}
\title{ABDS: Tool Suite for Analyzing Biologically Diverse Samples}
\description{
We developed the ABDS tool suite specifically for analyzing biologically diverse samples.
Mechanism-integrated group-wise pre-imputation is designed to recruit signature genes that represent informative missingness,
a cosine-based one-sample test is extended to detect down-regulated signature genes, and a unified heatmap is designed to
display multiple phenotypic groups. We discuss the methodological principles and demonstrate the effectiveness of three
analytics tools, supported by comparative evaluations and biomedical showcases.
}
