\name{abundance_matrix}
\alias{abundance_matrix}
\docType{data}
\title{abundance_matrix dataset}
\description{
  A dataset from the \pkg{abds} package. See \code{help(package = "abds")} for details.
}
\usage{abundance_matrix}
\format{A data frame.}
\source{Synthetic data. See \code{data-raw/generate_all.R}.}
\keyword{datasets}
