% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper functions.R
\name{cos_iCEG}
\alias{cos_iCEG}
\title{Calculate the cosine between data and CEG reference, e.g. c(1, 1, 1)}
\usage{
cos_iCEG(data)
}
\description{
Calculate the cosine between data and CEG reference, e.g. c(1, 1, 1)
}
