% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper functions.R
\name{cos_iDEG}
\alias{cos_iDEG}
\title{Calculate the cosine between data and closest iDEG reference, e.g. c(1, 0, 0)}
\usage{
cos_iDEG(data)
}
\description{
Calculate the cosine between data and closest iDEG reference, e.g. c(1, 0, 0)
}
