% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCOT.R
\name{cos_sDEG}
\alias{cos_sDEG}
\title{Calculate the cosine between a data matrix (features on rows and samples on columns)}
\usage{
cos_sDEG(data)
}
\description{
Calculate the cosine between a data matrix (features on rows and samples on columns)
}
