% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper functions.R
\name{cotMG}
\alias{cotMG}
\title{This function computes cosine values to select markers.}
\usage{
cotMG(
  data = NULL,
  Sest,
  thres.low = 0.05,
  thres.high = 1,
  cos.thres = 1,
  top = NULL,
  per = NULL
)
}
\description{
This function computes cosine values to select markers.
}
