% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCOT.R
\name{eCOT}
\alias{eCOT}
\title{Extended Cosine-based One-sample Test}
\usage{
eCOT(input, cos.thres = NULL, top = NULL, per = NULL)
}
\arguments{
\item{input}{The input matrix with features on rows and samples on columns. The input matrix should have only one sample in each 
group, or use 'super_sample' function first to create a super-sample that only has one sample in each group.}

\item{cos.thres}{The cosine threshold for DSGs.}

\item{top}{The upper bound of the number of total DSGs.}

\item{per}{The upper bound of the number of DSGs of each group.}
}
\description{
This function computes cosine values to select down-regulated signature genes (DSGs).
}
\examples{
output<-eCOT(input)
output<-eCOT(input,cos.thres=0.99)
output<-eCOT(input,top=100)
output<-eCOT(input,per=20)
# If you have multiple samples in one group, please use the 'super_sample' function to shrink the group size to one sample per group.
# See the end of this script.
# Please use only one of the filtering methods ('cos.thres', 'top', 'per') at one time.
}
