\name{env_regression}
\alias{env_regression}
\docType{data}
\title{env_regression dataset}
\description{
  A dataset from the \pkg{abds} package. See \code{help(package = "abds")} for details.
}
\usage{env_regression}
\format{A data frame.}
\source{Synthetic data. See \code{data-raw/generate_all.R}.}
\keyword{datasets}
