% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniHM.R
\name{heatmap_data_modification}
\alias{heatmap_data_modification}
\title{Function to modify data as per new heatmap design}
\usage{
heatmap_data_modification(tot_sg1)
}
\arguments{
\item{tot_sg1}{A numeric matrix containing the input data to be modified.}
}
\value{
A modified numeric matrix containing the scaled, log-transformed, and centered data.
}
\description{
This function modifies the input data for visualization in a heatmap by performing scaling, log transformation and centering.
}
