% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCOT.R, R/helper functions.R
\name{mycosine}
\alias{mycosine}
\title{Calculate the cosine between vector A and vector B}
\usage{
mycosine(A, B)

mycosine(A, B)
}
\description{
Calculate the cosine between vector A and vector B

Calculate the cosine between vector A and vector B
}
