\name{pa_matrix}
\alias{pa_matrix}
\docType{data}
\title{pa_matrix dataset}
\description{
  A dataset from the \pkg{abds} package. See \code{help(package = "abds")} for details.
}
\usage{pa_matrix}
\format{A data frame.}
\source{Synthetic data. See \code{data-raw/generate_all.R}.}
\keyword{datasets}
