% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniHM.R
\name{redblue}
\alias{redblue}
\title{Generate a colormap with shades of red and blue colors}
\usage{
redblue(m = nrow(colormap()))
}
\arguments{
\item{m}{an integer, number of colors in the colormap}
}
\value{
a matrix with three columns (r, g, b) representing the color map
}
\description{
The colormap is used to create a gradient from the blue color to the red color.
The function takes an optional argument m which specifies the number of colors in the colormap.
}
