% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_eCOT}
\alias{sample_data_eCOT}
\title{eCOT Sample Data}
\format{
\subsection{\code{sample_data_eCOT}}{

A data frame with 2753 rows and 3 columns:
}
}
\usage{
sample_data_eCOT
}
\description{
Sample data generated by Dirichlet distribution, with added signature genes (SGs) and down-regulated signature genes (DSGs).
}
\keyword{datasets}
