% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_MGpI}
\alias{sample_data_MGpI}
\title{MGpI Sample Data}
\format{
\subsection{\code{sample_data_MGpI}}{

A data frame with 29,772 rows and 16 columns:
}
}
\usage{
sample_data_MGpI
}
\description{
Sample data extracted from GSE19380 using only 16 pure samples from 4 groups (Primary neuronal culture,
Primary astrocytic culture, Primary oligodendrocytic culture, Primary microglial culture) according to the phenotypic data.
Then missing values are manually introduced by two mechanisms: lower limit of detection (LLOD) and missing at random (MAR),
with totally ~80,000 missing values and ~16,000 (20\%) missing due to MAR. Features with all values missing are removed.
}
\keyword{datasets}
