% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_uniHM}
\alias{sample_data_uniHM}
\title{uniHM Sample Data}
\format{
\subsection{\code{sample_data_uniHM}}{

A data frame with 12 rows and 10 columns:
}
}
\usage{
sample_data_uniHM
}
\description{
Sample data for illustrating a new heatmap design that can display both the differential pattern and
referenced quality of signature genes/down-regulated signature genes.
}
\keyword{datasets}
