% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{A function for illustrating data points in scatter plot with optional marker color-coding}
\usage{
scatter_plot(
  data,
  mg.info = FALSE,
  mg = NULL,
  mg.col = c("red", "orange", "dodgerblue", "green4", "purple")
)
}
\arguments{
\item{data}{With samples on the columns and features on the rows.}

\item{mg.info}{Do you have the info for marker genes?}

\item{mg}{A list of marker genes}

\item{mg.col}{A vector of colors}
}
\description{
A function for illustrating data points in scatter plot with optional marker color-coding
}
