\name{species_records}
\alias{species_records}
\docType{data}
\title{species_records dataset}
\description{
  A dataset from the \pkg{abds} package. See \code{help(package = "abds")} for details.
}
\usage{species_records}
\format{A data frame.}
\source{Synthetic data. See \code{data-raw/generate_all.R}.}
\keyword{datasets}
