% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCOT.R
\name{super_sample}
\alias{super_sample}
\title{Super-sample, if you have several samples in one group & reduce them into one sample per group.}
\usage{
super_sample(input, nRep)
}
\arguments{
\item{nRep}{Number of samples in each group. E.g., Group 1 has 4 samples, Group 2 has 3 samples, Group 3 has 3 samples.
Then nRep = c(4,3,3)}

\item{Input}{Input matrix, with features on rows and samples on columns.}
}
\description{
Super-sample, if you have several samples in one group & reduce them into one sample per group.
}
