\name{survey_multi}
\alias{survey_multi}
\docType{data}
\title{survey_multi dataset}
\description{
  A dataset from the \pkg{abds} package. See \code{help(package = "abds")} for details.
}
\usage{survey_multi}
\format{A data frame.}
\source{Synthetic data. See \code{data-raw/generate_all.R}.}
\keyword{datasets}
