\name{trap_messy}
\alias{trap_messy}
\docType{data}
\title{trap_messy dataset}
\description{
  A dataset from the \pkg{abds} package. See \code{help(package = "abds")} for details.
}
\usage{trap_messy}
\format{A data frame.}
\source{Synthetic data. See \code{data-raw/generate_all.R}.}
\keyword{datasets}
