% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniHM.R
\name{uniHM}
\alias{uniHM}
\alias{redblue}
\title{Unified Heatmap Design}
\usage{
redblue(m = nrow(colormap()))
}
\arguments{
\item{m}{an integer, number of colors in the colormap}
}
\value{
a matrix with three columns (r, g, b) representing the color map
}
\description{
To generate a colormap with shades of red and blue colors. 
The colormap is used to create a gradient from the blue color to the red color. 
The function takes an optional argument m which specifies the number of colors in the colormap.
}
\examples{
redblue()

}
